application/requirements.txt
google_api_python_client==2.141.0
google-cloud-pubsub==2.23.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
google-cloud-bigtable==2.26.0
google-cloud-asset==3.26.3
google-cloud-compute==1.19.2
protobuf==5.27.3
PyYAML==6.0.2
